Added in 2 dialogs. One being the zoom window and the other being the lua console

+++ Include/UI/EditMenu/OpenBoundaryDialog.h
@@ -29,7 +29,7 @@ public:
     
     ~openBoundaryDialog();
     
-    void getParameters(int &numberLayers, double &radius, double &horizontalCenter, double &verticalCenter, OpenBoundaryEdge &edgeType);
+    void getParameters(long &numberLayers, double &radius, double &horizontalCenter, double &verticalCenter, OpenBoundaryEdge &edgeType);
 };
